Created files computeDescriptors.R and removeConstantDesc.R
The two files contain functions that can be executed to computeDescriptors and remove constant columns. It is done to make the code more modulare and easier to comprehend. Descriptors are computed using the rcdk package. Constant Descriptors are removed because it is assumed that they do not add predicitve information concerning the boilingpoint when they do not change with differences in the molecules.