Added description for execution of code and recommendation for rstudio
Link to rstudio was added togehter with a recommendation to install it. Important notes on how to execute the code and information about randomization of data as well as selection of chemical descriptors was added to the section execution of code. These principles will guarantee better usability and reproducibility of the code.